Create maps activity

+++ app/src/main/java/mperezf/mimo/gruposcaminosantiago/data/model/MessageData.kt
@@ -4,5 +4,5 @@ data class MessageData(
     val id: Int?,
     val author: UserData?,
     val content: String?,
-    val whenSent: Int?
+    val whenSent: Long?
 )
